% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHMMpackage.R
\docType{data}
\name{accelero}
\alias{accelero}
\title{Accelerometer data}
\description{
Accelerometer data measured each 5 minutes on three subjects
}
\examples{
  data(accelero)
}
\references{
Huang, Q., Cohen, D., Komarzynski, S., Li, X.-M., Innominato, P., Lévi, F., and Finkenstädt, B. (2018b). Hidden markov models for monitoring circadian rhythmicity in telemetric activity data. Journal of The Royal Society Interface, 15(139):20170885
}
\keyword{datasets}
