% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\docType{class}
\name{mhmmparam-class}
\alias{mhmmparam-class}
\title{Constructor of \code{\linkS4class{mhmmparam}} class}
\description{
\describe{
  \item{K}{numeric. Number of classes.}
  \item{M}{numeric. Number of latent states (activity levels).}
  \item{A}{list. Matrices of the transition probailities for each class.}
  \item{delta}{numeric. Proportions of the classes.}
  \item{pi}{matrix. Probabilities of the latent states per class (stationary distribution of the Markov chains).}
  \item{lambda}{list. Parameters of the emission laws (eps: proportions of the zero-inflated component, a: shapes of the gamma distributions, b: rates of the gamma distributions)}
}
}
\examples{
  getSlots("mhmmparam")

}
