% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\docType{class}
\name{mhmmresults-class}
\alias{mhmmresults-class}
\title{Constructor of \code{\linkS4class{mhmmresults}} class}
\description{
\describe{
  \item{param}{\linkS4class{mhmmparam}. MLE of the model parameters}
  \item{data}{\linkS4class{mhmmdata}. Data}
  \item{partitions}{list. Elements of the latent variables (partition among subjects and latent states)}
  \item{probabilities}{list. Posterior probabilities of the latent variables.}
  \item{meantimesperstates}{matrix. Mean time spent by each subject (row) into each activity level (column).}
  \item{meanvalueperstates}{numeric. Summary statistics of the states.}
  \item{loglike}{numeric. Loglikelihood.}
  \item{bic}{numeric. BIC.}
}
}
\examples{
  getSlots("mhmmresults")

}
